💄 Suggestion working, WIP
