--- frontend/models/feedback-task-model/record-suggestion/recordSuggestion.queries.js
@@ -1,7 +1,0 @@
-import { RecordSuggestion as RecordSuggestionModel } from "./RecordSuggestion.model";
-
-// UPSERT
-const upsertRecordSuggestions = (recordSuggestions) =>
-  RecordSuggestionModel.insertOrUpdate({ data: recordSuggestions });
-
-export { upsertRecordSuggestions };
